Added `tokenize` bypass stripping patterns
We found that emails in arrows were getting stripped from strings,
and this is bad, because it's a very common text.

eg: "Alan Blount <alan@zeroasterisk.com>, Someone <some@one.com>"

It was stripping off the email and returning just:

eg: "Alan Blount , Someone "

Obviously, this isn't what we wanted...

----

So I implemented a pattern for tokeninzing portions of the input string
before cleaning it.

And I defaulted a pattern for `emailInArrows` into `$config.tokenizations`.

And I defaulted it to be enabled for all keys, in `$config.sanitizationKeyMap.$key.tokenize`.

----

This pattern can easily be extened / overridden by anyone's config.
(though any existing config may need to be modified to include it now)

Any other "known" patterns we should default on?  Open an issue or pull request please!